Re-org device automations
