+++ homeassistant/components/device_automation/trigger.py
@@ -1,5 +1,5 @@
 """Offer device oriented automation."""
-from typing import cast
+from typing import Protocol, cast
 
 import voluptuous as vol
 
@@ -21,17 +21,41 @@ from .exceptions import InvalidDeviceAutomationConfig
 TRIGGER_SCHEMA = DEVICE_TRIGGER_BASE_SCHEMA.extend({}, extra=vol.ALLOW_EXTRA)
 
 
+class DeviceAutomationTriggerProtocol(Protocol):
+    """Define the format of device_trigger modules.
+
+    Each module must define either TRIGGER_SCHEMA or async_validate_trigger_config.
+    """
+
+    TRIGGER_SCHEMA: vol.Schema
+
+    async def async_validate_trigger_config(
+        self, hass: HomeAssistant, config: ConfigType
+    ) -> ConfigType:
+        """Validate config."""
+        raise NotImplementedError
+
+    async def async_attach_trigger(
+        self,
+        hass: HomeAssistant,
+        config: ConfigType,
+        action: AutomationActionType,
+        automation_info: AutomationTriggerInfo,
+    ) -> CALLBACK_TYPE:
+        """Attach a trigger."""
+        raise NotImplementedError
+
+
 async def async_validate_trigger_config(
     hass: HomeAssistant, config: ConfigType
 ) -> ConfigType:
     """Validate config."""
-    platform = await async_get_device_automation_platform(
-        hass, config[CONF_DOMAIN], DeviceAutomationType.TRIGGER
-    )
-    if not hasattr(platform, "async_validate_trigger_config"):
-        return cast(ConfigType, platform.TRIGGER_SCHEMA(config))
-
     try:
+        platform = await async_get_device_automation_platform(
+            hass, config[CONF_DOMAIN], DeviceAutomationType.TRIGGER
+        )
+        if not hasattr(platform, "async_validate_trigger_config"):
+            return cast(ConfigType, platform.TRIGGER_SCHEMA(config))
         return await platform.async_validate_trigger_config(hass, config)
     except InvalidDeviceAutomationConfig as err:
         raise vol.Invalid(str(err) or "Invalid trigger configuration") from err
