+++ homeassistant/helpers/condition.py
@@ -13,10 +13,7 @@ import sys
 from typing import Any, cast
 
 from homeassistant.components import zone as zone_cmp
-from homeassistant.components.device_automation import (
-    DeviceAutomationType,
-    async_get_device_automation_platform,
-)
+from homeassistant.components.device_automation import condition as device_condition
 from homeassistant.components.sensor import SensorDeviceClass
 from homeassistant.const import (
     ATTR_DEVICE_CLASS,
@@ -30,7 +27,6 @@ from homeassistant.const import (
     CONF_BELOW,
     CONF_CONDITION,
     CONF_DEVICE_ID,
-    CONF_DOMAIN,
     CONF_ENTITY_ID,
     CONF_ID,
     CONF_STATE,
@@ -872,10 +868,8 @@ async def async_device_from_config(
     hass: HomeAssistant, config: ConfigType
 ) -> ConditionCheckerType:
     """Test a device condition."""
-    platform = await async_get_device_automation_platform(
-        hass, config[CONF_DOMAIN], DeviceAutomationType.CONDITION
-    )
-    return trace_condition_function(platform.async_condition_from_config(hass, config))
+    checker = await device_condition.async_condition_from_config(hass, config)
+    return trace_condition_function(checker)
 
 
 async def async_trigger_from_config(
@@ -931,15 +925,10 @@ async def async_validate_condition_config(
             sub_cond = await async_validate_condition_config(hass, sub_cond)
             conditions.append(sub_cond)
         config["conditions"] = conditions
+        return config
 
     if condition == "device":
-        config = cv.DEVICE_CONDITION_SCHEMA(config)
-        platform = await async_get_device_automation_platform(
-            hass, config[CONF_DOMAIN], DeviceAutomationType.CONDITION
-        )
-        if hasattr(platform, "async_validate_condition_config"):
-            return await platform.async_validate_condition_config(hass, config)
-        return cast(ConfigType, platform.CONDITION_SCHEMA(config))
+        return await device_condition.async_validate_condition_config(hass, config)
 
     if condition in ("numeric_state", "state"):
         validator = cast(
